backend/requirements.txt
Django==3.2.3
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djoser==2.2.0
gunicorn==20.1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
